Refactor download and install for languages into their own package
